Revert changes to wobblyedge
