+++ content/webapp/views/components/WShape/index.tsx
@@ -1,19 +1,27 @@
+import { CSSProperties } from 'react';
+
 import { PaletteColor, themeValues } from '@weco/common/views/themes/config';
 
 // eslint-data-component: intentionally omitted
 const WShape = ({
   variant,
   color,
+  styles,
 }: {
   variant: 'full-1' | 'full-2' | 'full-3' | 'edge-1' | 'edge-2';
   color?: PaletteColor;
+  styles?: CSSProperties;
 }) => {
   const fillColor = color ? themeValues.color(color) : 'currentColor';
 
   switch (variant) {
     case 'full-1':
       return (
-        <svg viewBox="0 0 1675 972" xmlns="http://www.w3.org/2000/svg">
+        <svg
+          viewBox="0 0 1675 972"
+          xmlns="http://www.w3.org/2000/svg"
+          {...(styles && { style: styles })}
+        >
           <path
             d="M1605.87 932.699C1605.71 934.448 1605.47 936.221 1604.65 937.779C1600.11 946.367 1585.42 951.669 1569.78 949.899C1549.24 947.573 1523.28 944.587 1508.01 942.652C1479.87 939.088 1440.11 933.917 1417.92 931.07C1395.73 928.223 1381.76 926.489 1374.87 925.601C1367.98 924.714 1368.55 924.479 1367.01 929.193C1365.47 933.906 1361.69 946.058 1359.33 953.199L1357.62 957.905C1354.01 967.797 1337.59 973.849 1320.64 971.491C1283 966.251 1219.48 957.363 1183.21 952.02C1122.85 943.13 1062.46 934.344 1002.12 925.394C717.979 883.246 450.955 840.102 165.718 801.165C150.591 799.1 119.769 793.678 107.708 790.855C95.6475 788.032 71.4638 786.852 66.0893 743.938C60.7148 701.025 15.2205 239.679 9.4345 196.025C3.64852 152.371 1.84585 139.06 0.883707 131.762C-0.0784382 124.463 -1.09053 125.331 10.4133 124.14C21.9172 122.949 38.9793 121.594 58.5415 119.564C78.1038 117.534 102.814 114.63 127.825 112.066C464.318 77.5748 800.851 43.1769 1137.34 8.68761C1165.35 5.81669 1190.39 2.9653 1207.82 1.00231C1225.24 -0.960688 1233.31 -1.38181 1233.74 12.7185C1234.32 31.6684 1233.68 23.9893 1233.8 33.3166C1233.92 42.644 1234.36 66.8205 1234.5 71.3213C1234.63 75.8221 1236.98 77.1498 1243.77 76.3489L1481.15 54.8696C1504.57 52.7035 1514.24 73.8692 1515.99 84.2233C1517.81 95.0028 1520.1 108.153 1521.87 117.199C1524.8 132.123 1529.86 157.597 1531.44 165.58C1531.72 166.978 1533.7 168.041 1536.07 168.034C1556.9 167.97 1649.7 167.686 1662.72 167.6C1677.4 167.503 1674.68 167.274 1672.92 174.077L1605.87 932.699Z"
             fill={fillColor}
@@ -22,7 +30,11 @@ const WShape = ({
       );
     case 'full-2':
       return (
-        <svg viewBox="0 0 1010 849" xmlns="http://www.w3.org/2000/svg">
+        <svg
+          viewBox="0 0 1010 849"
+          xmlns="http://www.w3.org/2000/svg"
+          {...(styles && { style: styles })}
+        >
           <path
             d="M640.04 58.0884C632.254 72.0628 624.75 85.4953 617.779 97.9745C594.192 83.9858 571.519 69.9552 548.296 56.9291C524.766 43.7309 498.931 36.8494 472.531 32.5497C464.571 31.2531 456.108 32.4772 447.947 33.3169C446.002 33.517 443.796 36.5141 442.634 38.722C421.629 78.6272 400.624 118.534 380.014 158.648C376.578 165.335 373.851 166.166 367.727 162.005C340.815 143.717 313.604 125.88 286.533 107.832C269.904 96.7463 253.314 85.5994 236.742 74.4269C230.47 70.1986 222.993 72.1447 219.471 79.1737C175.517 166.909 131.568 254.647 87.6452 342.399C59.0418 399.545 30.5672 456.756 1.7806 513.807C-1.26965 519.852 -0.541556 522.774 5.05805 526.579C51.2255 557.954 97.1346 589.72 143.16 621.309C224.451 677.103 305.754 732.878 387.088 788.607C404.394 800.465 421.314 813.039 439.368 823.609C469.993 841.538 502.737 852.957 538.997 847.724C547.495 846.497 555.703 842.712 563.869 839.617C565.929 838.837 567.679 836.415 568.952 834.354C596.281 790.127 623.638 745.917 650.634 701.482C654.096 695.783 657.015 695.133 662.471 697.987C678.419 706.327 694.703 713.969 712.75 716.526C728.704 718.787 744.604 719.172 759.751 712.121C762.294 710.938 764.842 708.789 766.356 706.412C846.038 581.35 925.61 456.217 1005.18 331.083C1006.81 328.521 1008.29 325.863 1010 322.979C974.884 289.121 939.652 255.531 904.841 221.498C855.651 173.406 806.781 124.978 757.756 76.7122C732.882 52.2229 707.97 27.7736 683.075 3.30649C677.862 -1.81711 672.42 -0.927186 668.863 5.56565C659.366 22.9014 649.839 40.2206 640.04 58.0884Z"
             fill={fillColor}
@@ -31,7 +43,11 @@ const WShape = ({
       );
     case 'full-3':
       return (
-        <svg viewBox="0 0 870 822" xmlns="http://www.w3.org/2000/svg">
+        <svg
+          viewBox="0 0 870 822"
+          xmlns="http://www.w3.org/2000/svg"
+          {...(styles && { style: styles })}
+        >
           <path
             d="M775.84 265.418C760.87 270.412 746.466 275.2 733.084 279.648C725.178 253.769 717.968 228.467 709.688 203.531C701.298 178.265 687.028 156.065 670.629 135.369C665.684 129.128 658.697 124.388 652.186 119.57C650.634 118.421 647.032 119.136 644.717 119.97C602.865 135.051 561.011 150.133 519.308 165.628C512.356 168.211 509.814 167.001 508.113 159.897C500.64 128.675 492.65 97.5826 484.901 66.4291C480.142 47.2927 475.451 28.1382 470.79 8.97648C469.026 1.72469 462.278 -1.83321 455.036 0.941876C364.644 35.5798 274.254 70.2232 183.875 104.894C125.018 127.472 66.2105 150.184 7.28278 172.571C1.03877 174.943 -0.372427 177.555 1.18225 184.051C14.0001 237.614 26.3701 291.29 38.9421 344.915C61.1468 439.627 83.3719 534.335 105.651 629.029C110.392 649.178 114.374 669.591 120.517 689.3C130.936 722.731 147.23 752.832 177.159 773.147C184.173 777.909 192.678 780.613 200.692 783.794C202.713 784.596 205.6 783.996 207.899 783.343C257.237 769.323 306.585 755.335 355.819 740.942C362.134 739.096 364.695 740.565 366.778 746.273C372.866 762.957 379.664 779.356 391.128 793.225C401.263 805.487 412.606 816.346 428.342 821.291C430.984 822.121 434.271 822.251 436.956 821.528C578.225 783.484 719.462 745.314 860.698 707.143C863.589 706.361 866.437 705.41 869.601 704.448C866.495 656.43 863.127 608.531 860.36 560.588C856.451 492.842 852.999 425.064 849.327 357.301C847.464 322.92 845.546 288.542 843.653 254.162C843.256 246.963 838.695 243.99 831.784 246.351C813.331 252.655 794.867 258.927 775.84 265.418Z"
             fill={fillColor}
@@ -40,7 +56,11 @@ const WShape = ({
       );
     case 'edge-1':
       return (
-        <svg viewBox="0 0 1256 127" xmlns="http://www.w3.org/2000/svg">
+        <svg
+          viewBox="0 0 1256 127"
+          xmlns="http://www.w3.org/2000/svg"
+          {...(styles && { style: styles })}
+        >
           <path
             d="M961.015 1.15779C977.287 -0.546414 985.386 -2.49896 985.386 12.2896C985.386 30.0005 985.342 22.8145 985.386 31.525C985.429 40.2357 985.295 62.8153 985.386 67.0191C985.477 71.2226 987.661 72.4794 994.005 71.7818L1211.63 53.4791C1233.5 51.6294 1239.94 66.2941 1243.92 81.147C1247.9 96 1256 127.001 1256 127.001H0V103.325C0 97.1348 4.70796 91.9606 10.8702 91.3779C305.641 63.5039 600.44 35.6958 895.205 7.81404C921.359 5.34016 944.741 2.86205 961.015 1.15779Z"
             fill={fillColor}
@@ -49,7 +69,11 @@ const WShape = ({
       );
     case 'edge-2':
       return (
-        <svg viewBox="0 0 1440 68" xmlns="http://www.w3.org/2000/svg">
+        <svg
+          viewBox="0 0 1440 68"
+          xmlns="http://www.w3.org/2000/svg"
+          {...(styles && { style: styles })}
+        >
           <path
             d="M628.053 0.214678C647.505 -0.197297 666.972 0.112137 686.651 0.112139C688.428 13.2288 690.126 25.7594 691.856 38.534C708.421 37.7375 723.913 36.8664 739.416 36.2713C780.801 34.6829 822.189 33.1527 863.58 31.7196C886.421 30.9287 909.272 30.4219 932.116 29.7059C946.42 29.2575 960.72 28.675 975.021 28.1239C1022.17 26.3069 1069.31 24.3851 1116.46 22.7225C1127.14 22.3462 1126.91 22.7729 1127.76 33.5506C1128.59 43.9339 1129.99 54.2714 1131.23 65.4744C1164.87 64.2622 1197.47 62.9922 1230.07 61.9383C1267.22 60.7377 1304.37 59.7875 1341.52 58.619C1367.61 57.7982 1393.7 56.6414 1419.79 55.9784C1426.39 55.8106 1433.03 56.9561 1440 57.5272V68.0008H0V27.2186C34.3211 25.754 68.6363 24.1519 102.957 22.6756C151.356 20.5937 199.758 18.5894 248.157 16.5155C281.758 15.0757 315.354 13.5488 348.954 12.1063C397.102 10.0392 445.249 7.95668 493.401 5.99105C523.754 4.75203 554.119 3.80369 584.472 2.54866C599.006 1.94766 613.516 0.522568 628.053 0.214678Z"
             fill={fillColor}
